Reorganization and junk (#15)
* Extract tests module

* extract lib.rs to more modules

* add license to tests.rs

* reorganize

* add is_empty function to Page struct to match len()

* update gitignore

* remove .idea folder from repo
