+++ .gitignore
@@ -1,2 +1,5 @@
 /target
 Cargo.lock
+/.idea
+.DS_Store
+/**/.DS_Store
