+++ src/tests.rs
@@ -1,0 +1,15 @@
+// Sprengui
+// Copyright © 2020-2021 Jeron Aldaron Lau.
+//
+// Licensed under any of:
+//  - Apache License, Version 2.0 (https://www.apache.org/licenses/LICENSE-2.0)
+//  - MIT License (https://mit-license.org/)
+//  - Boost Software License, Version 1.0 (https://www.boost.org/LICENSE_1_0.txt)
+// At your choosing (See accompanying files LICENSE_APACHE_2_0.txt,
+// LICENSE_MIT.txt and LICENSE_BOOST_1_0.txt).
+
+// TODO: add real tests
+#[test]
+fn it_works() {
+    assert_eq!(2 + 2, 4);
+}
